+++ lua/plugins/config/golang_enhancement.lua
@@ -20,4 +20,52 @@ return function(use)
 			ft = { 'go' },
 		})
 	end
+
+	use({
+		'rafaelsq/nvim-goc.lua',
+		config = function()
+			require('nvim-goc').setup({})
+
+			require('which-key').register({
+				['<leader>c'] = {
+					name = '+coverage',
+					c = {
+						function()
+							require('nvim-goc').ClearCoverage()
+						end,
+						'Clear Coverage',
+					},
+					r = {
+						function()
+							require('nvim-goc').Coverage()
+						end,
+						'Coverage',
+					},
+					t = {
+						function()
+							require('nvim-goc').CoverageFunc()
+						end,
+						'Coverage Func',
+					},
+				},
+			})
+		end,
+		ft = { 'go' },
+	})
+
+	use({
+		'edolphin-ydf/goimpl.nvim',
+		config = function()
+			local ok, telescope = pcall(require, 'telescope')
+			if not ok then
+				return
+			end
+
+			telescope.load_extension('goimpl')
+			vim.cmd(
+				[[au filetype go lua require('which-key').register({qi = {function() require('telescope').extensions.goimpl.goimpl() end, 'Go Impl'}})]]
+			)
+		end,
+		ft = { 'go' },
+	})
 end